Extracted common part of pruning
