+++ nncf/common/pruning/export_helpers.py
@@ -1,0 +1,37 @@
+"""
+ Copyright (c) 2020 Intel Corporation
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+"""
+
+from nncf.common.graph.graph import NNCFNode
+
+
+class DefaultMetaOp:
+    subtypes = []
+    additional_types = []
+
+    @classmethod
+    def accept_pruned_input(cls, node: NNCFNode):
+        """
+        :return: accept_pruned_input - can this operation work with pruned input or not
+        """
+        raise NotImplementedError
+
+    @classmethod
+    def get_all_op_aliases(cls):
+        """
+        :return: list of all aliases of types in metatype
+        """
+        op_types = []
+        for subtype in cls.subtypes:
+            op_types.extend(subtype.get_all_aliases())
+        op_types = list(set(op_types)) + cls.additional_types
+        return op_types
