+++ nncf/common/pruning/pruning_node_selector.py
@@ -11,14 +11,24 @@
  limitations under the License.
 """
 
-from typing import Dict, List, Optional
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Tuple
 from collections import defaultdict
 
-from nncf.pruning.utils import get_sources_of_node, is_depthwise_conv, get_previous_conv, get_first_pruned_nodes, \
-    get_last_pruned_nodes, is_grouped_conv, is_conv_with_downsampling
+from nncf.common.graph.graph import NNCFGraph
+from nncf.common.graph.graph import NNCFNode
+from nncf.common.pruning.utils import get_sources_of_node
+from nncf.common.pruning.utils import get_first_pruned_nodes
+from nncf.common.pruning.utils import get_last_pruned_nodes
+from nncf.common.pruning.utils import get_previous_conv
+from nncf.common.pruning.utils import is_grouped_conv
 from nncf.pruning.export_helpers import PruningOperationsMetatypeRegistry
-from nncf.pruning.model_analysis import ModelAnalyzer, Clusterization, cluster_special_ops, NodesCluster
-from nncf.dynamic_graph.graph import NNCFGraph, NNCFNode
+from nncf.common.pruning.model_analysis import ModelAnalyzer
+from nncf.common.pruning.model_analysis import Clusterization
+from nncf.common.pruning.model_analysis import cluster_special_ops
+from nncf.common.pruning.model_analysis import NodesCluster
 from nncf.common.utils.logger import logger as nncf_logger
 from nncf.utils import should_consider_scope
 
@@ -54,6 +64,8 @@ class PruningNodeSelector:
         self._pruning_operator_metatypes = pruning_operator_metatypes
         pruning_op_metatypes_dict = self._pruning_operator_metatypes.registry_dict
         self._identity_mask_propagation_op_metatype = pruning_op_metatypes_dict['identity_mask_propagation']
+        self._stop_propagation_op_metatype = pruning_op_metatypes_dict['stop_propagation_ops']
+        self._convolution_op_metatype = pruning_op_metatypes_dict['convolution']
 
         self._prune_operations = prune_operations
         self._grouping_operations = grouping_operations
@@ -126,8 +138,9 @@ class PruningNodeSelector:
         for node in all_nodes_to_prune:
             cluster_id = pruned_nodes_clusterization.get_cluster_by_node_id(node.node_id).id
 
-            if is_depthwise_conv(node):
-                previous_conv = get_previous_conv(graph, node)
+            if self._is_depthwise_conv(node):
+                stop_propagation_ops = self._stop_propagation_op_metatype.get_all_op_aliases()
+                previous_conv = get_previous_conv(graph, node, self._prune_operations, stop_propagation_ops)
                 if previous_conv:
                     previous_conv_cluster_id = pruned_nodes_clusterization.get_cluster_by_node_id(
                         previous_conv.node_id).id
@@ -148,7 +161,7 @@ class PruningNodeSelector:
             pruned_nodes_clusterization.merge_list_of_clusters(previous_clusters)
 
         # 6. Checks for groups (all nodes in group can be pruned or all group can't be pruned).
-        model_analyser = ModelAnalyzer(graph, self._pruning_operator_metatypes)
+        model_analyser = ModelAnalyzer(graph, self._pruning_operator_metatypes, self._is_depthwise_conv)
         can_prune_analysis = model_analyser.analyse_model_before_pruning()
         self._check_pruning_groups(graph, pruned_nodes_clusterization, can_prune_analysis)
         return pruned_nodes_clusterization
@@ -160,8 +173,9 @@ class PruningNodeSelector:
         """
         ret = defaultdict(list)
         for node in graph.get_nodes_by_types(self._prune_operations):
-            ret[node.op_exec_context.scope_in_model].append(node.node_id)
-        return [ret[scope] for scope in ret if len(ret[scope]) > 1]
+            module_identifier = self._get_module_identifier(node)
+            ret[module_identifier].append(node.node_id)
+        return [ret[module_identifier] for module_identifier in ret if len(ret[module_identifier]) > 1]
 
     def _check_pruning_groups(self, graph: NNCFGraph, pruned_nodes_clusterization: Clusterization,
                               can_prune: Dict[str, bool]):
@@ -173,7 +187,7 @@ class PruningNodeSelector:
         :param can_prune: dict with bool value: can this node be pruned or not
         """
         for cluster in pruned_nodes_clusterization.get_all_clusters():
-            cluster_nodes_names = [str(n.op_exec_context.scope_in_model) for n in cluster.nodes]
+            cluster_nodes_names = [self._get_module_identifier(n) for n in cluster.nodes]
             # Check whether this node should be pruned according to the user-defined algorithm constraints
             should_prune_nodes = [self._is_module_prunable(graph, node) for node in cluster.nodes]
 
@@ -186,7 +200,7 @@ class PruningNodeSelector:
                                                                             ", ".join(shouldnt_prune_msgs)))
                 pruned_nodes_clusterization.delete_cluster(cluster.id)
             elif not all(can_prune_nodes):
-                cant_prune_nodes_names = [str(node.op_exec_context.scope_in_model) for node in cluster.nodes
+                cant_prune_nodes_names = [self._get_module_identifier(node) for node in cluster.nodes
                                           if not can_prune[node.node_id]]
                 nncf_logger.info("Group of nodes [{}] can't be pruned, because {} nodes can't be pruned "
                                  "according to model analysis"
@@ -195,7 +209,7 @@ class PruningNodeSelector:
             else:
                 nncf_logger.info("Group of nodes [{}] will be pruned together.".format(", ".join(cluster_nodes_names)))
 
-    def _is_module_prunable(self, graph: NNCFGraph, node: NNCFNode):
+    def _is_module_prunable(self, graph: NNCFGraph, node: NNCFNode) -> Tuple[bool, str]:
         """
         Check whether we should prune module corresponding to provided node
         according to algorithm parameters.
@@ -210,25 +224,34 @@ class PruningNodeSelector:
 
         input_non_pruned_nodes = get_first_pruned_nodes(graph, self._prune_operations + ['linear'])
         output_non_pruned_nodes = get_last_pruned_nodes(graph, self._prune_operations + ['linear'])
-        node_scope_str = str(node.op_exec_context.scope_in_model)
+        module_identifier = self._get_module_identifier(node)
 
-        if not should_consider_scope(node_scope_str, self._target_scopes, self._ignored_scopes):
-            msg = "Ignored adding Weight Pruner in scope: {}".format(node_scope_str)
+        if not should_consider_scope(module_identifier, self._target_scopes, self._ignored_scopes):
+            msg = 'Ignored adding Weight Pruner in: {}'.format(module_identifier)
             prune = False
         elif not self._prune_first and node in input_non_pruned_nodes:
-            msg = "Ignored adding Weight Pruner in scope: {} because"\
-                             " this scope is one of the first convolutions".format(node_scope_str)
+            msg = 'Ignored adding Weight Pruner in: {} because'\
+                             ' this scope is one of the first convolutions'.format(module_identifier)
             prune = False
         elif not self._prune_last and node in output_non_pruned_nodes:
-            msg = "Ignored adding Weight Pruner in scope: {} because"\
-                             " this scope is one of the last convolutions".format(node_scope_str)
+            msg = "Ignored adding Weight Pruner in: {} because"\
+                             " this scope is one of the last convolutions".format(module_identifier)
             prune = False
-        elif is_grouped_conv(node) and not is_depthwise_conv(node):
-            msg = "Ignored adding Weight Pruner in scope: {} because" \
-                  " this scope is grouped convolution".format(node_scope_str)
+        elif is_grouped_conv(node) and not self._is_depthwise_conv(node):
+            msg = 'Ignored adding Weight Pruner in: {} because' \
+                  ' this scope is grouped convolution'.format(module_identifier)
             prune = False
-        elif not self._prune_downsample_convs and is_conv_with_downsampling(node):
-            msg = "Ignored adding Weight Pruner in scope: {} because"\
-                             " this scope is convolution with downsample".format(node_scope_str)
+        elif not self._prune_downsample_convs and self._is_conv_with_downsampling(node):
+            msg = 'Ignored adding Weight Pruner in: {} because'\
+                             ' this scope is convolution with downsample'.format(module_identifier)
             prune = False
         return prune, msg
+
+    def _get_module_identifier(self, node: NNCFNode) -> str:
+        raise NotImplementedError
+
+    def _is_depthwise_conv(self, node: NNCFNode) -> bool:
+        raise NotImplementedError
+
+    def _is_conv_with_downsampling(self, node: NNCFNode) -> bool:
+        raise NotImplementedError
